# frontendnew/phonebook-frontend/src/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Jefferson Lab Phonebook",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/v1/board-members/": {
            "get": {
                "tags": [
                    "Board Members"
                ],
                "summary": "List all institutional/executive board memberships",
                "description": "Retrieves a list of all board memberships, with optional filtering by board type, member, or institution.",
                "operationId": "Board Members-list_board_memberships",
                "parameters": [
                    {
                        "name": "skip",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "title": "Skip"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 100,
                            "title": "Limit"
                        }
                    },
                    {
                        "name": "board_type",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "$ref": "#/components/schemas/BoardType"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Board Type"
                        }
                    },
                    {
                        "name": "member_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "integer"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Member Id"
                        }
                    },
                    {
                        "name": "institution_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "integer"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Institution Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/InstitutionalBoardMemberResponse"
                                    },
                                    "title": "Response Board Members-List Board Memberships"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Board Members"
                ],
                "summary": "Create a new institutional/executive board membership",
                "description": "Creates a new board membership for a member at an institution.",
                "operationId": "Board Members-create_board_membership",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/InstitutionalBoardMemberCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InstitutionalBoardMemberResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/board-members/{ibm_id}": {
            "get": {
                "tags": [
                    "Board Members"
                ],
                "summary": "Get board membership by ID",
                "description": "Retrieves a single institutional/executive board membership by its unique ID.",
                "operationId": "Board Members-get_board_membership",
                "parameters": [
                    {
                        "name": "ibm_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Ibm Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InstitutionalBoardMemberResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Board Members"
                ],
                "summary": "Update a board membership",
                "description": "Updates an existing institutional/executive board membership's details by its ID.",
                "operationId": "Board Members-update_board_membership",
                "parameters": [
                    {
                        "name": "ibm_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Ibm Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/InstitutionalBoardMemberUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InstitutionalBoardMemberResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Board Members"
                ],
                "summary": "Delete a board membership",
                "description": "Deletes an institutional/executive board membership by its ID.",
                "operationId": "Board Members-delete_board_membership",
                "parameters": [
                    {
                        "name": "ibm_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Ibm Id"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/groups/{group_id}/members": {
            "get": {
                "tags": [
                    "Working Groups"
                ],
                "summary": "List members of a specific working group",
                "description": "Retrieves a list of members belonging to a specific working group.",
                "operationId": "Working Groups-list_group_members_of_group",
                "parameters": [
                    {
                        "name": "group_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Group Id"
                        }
                    },
                    {
                        "name": "skip",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "title": "Skip"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 100,
                            "title": "Limit"
                        }
                    },
                    {
                        "name": "role",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "$ref": "#/components/schemas/GroupRole"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Role"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/GroupMemberResponse"
                                    },
                                    "title": "Response Working Groups-List Group Members Of Group"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Working Groups"
                ],
                "summary": "Add a member to a working group",
                "description": "Adds an existing member to a specific working group with a given role.",
                "operationId": "Working Groups-add_member_to_group",
                "parameters": [
                    {
                        "name": "group_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Group Id"
                        }
                    },
                    {
                        "name": "member_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Member Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GroupMemberCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GroupMemberResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "BoardType": {
                "type": "string",
                "enum": [
                    "institutional",
                    "executive"
                ],
                "title": "BoardType"
            },
...
            "GroupMemberCreate": {
                "properties": {
                    "group_id": {
                        "type": "integer",
                        "title": "Group Id"
                    },
                    "member_id": {
                        "type": "integer",
                        "title": "Member Id"
                    },
                    "role": {
                        "$ref": "#/components/schemas/GroupRole"
                    }
                },
                "type": "object",
                "required": [
                    "group_id",
                    "member_id",
                    "role"
                ],
                "title": "GroupMemberCreate"
            },
            "GroupMemberResponse": {
                "properties": {
                    "group_id": {
                        "type": "integer",
                        "title": "Group Id"
                    },
                    "member_id": {
                        "type": "integer",
                        "title": "Member Id"
                    },
                    "role": {
                        "$ref": "#/components/schemas/GroupRole"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    }
                },
                "type": "object",
                "required": [
                    "group_id",
                    "member_id",
                    "role",
                    "id"
                ],
                "title": "GroupMemberResponse"
            },
...
            "GroupRole": {
                "type": "string",
                "enum": [
                    "member",
                    "convenor",
                    "co-convenor"
                ],
                "title": "GroupRole"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "InstitutionalBoardMemberCreate": {
                "properties": {
                    "member_id": {
                        "type": "integer",
                        "title": "Member Id"
                    },
                    "institution_id": {
                        "type": "integer",
                        "title": "Institution Id"
                    },
                    "board_type": {
                        "$ref": "#/components/schemas/BoardType"
                    },
                    "role": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Role"
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date",
                        "title": "Start Date"
                    },
                    "end_date": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "End Date"
                    },
                    "is_chair": {
                        "type": "boolean",
                        "title": "Is Chair",
                        "default": false
                    }
                },
                "type": "object",
                "required": [
                    "member_id",
                    "institution_id",
                    "board_type",
                    "start_date"
                ],
                "title": "InstitutionalBoardMemberCreate"
            },
            "InstitutionalBoardMemberResponse": {
                "properties": {
                    "member_id": {
                        "type": "integer",
                        "title": "Member Id"
                    },
                    "institution_id": {
                        "type": "integer",
                        "title": "Institution Id"
                    },
                    "board_type": {
                        "$ref": "#/components/schemas/BoardType"
                    },
                    "role": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Role"
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date",
                        "title": "Start Date"
                    },
                    "end_date": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "End Date"
                    },
                    "is_chair": {
                        "type": "boolean",
                        "title": "Is Chair",
                        "default": false
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    }
                },
                "type": "object",
                "required": [
                    "member_id",
                    "institution_id",
                    "board_type",
                    "start_date",
                    "id"
                ],
                "title": "InstitutionalBoardMemberResponse"
            },
            "InstitutionalBoardMemberUpdate": {
                "properties": {
                    "member_id": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Member Id"
                    },
                    "institution_id": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Institution Id"
                    },
                    "board_type": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/BoardType"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "role": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Role"
                    },
                    "start_date": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Start Date"
                    },
                    "end_date": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "End Date"
                    },
                    "is_chair": {
                        "anyOf": [
                            {
                                "type": "boolean"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Is Chair"
                    }
                },
                "type": "object",
                "title": "InstitutionalBoardMemberUpdate"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}